지역 검색 안하고 다음 화면 넘어갔을 때 정보 상태 수정

+++ app/src/main/java/com/jccgs/travelplanner_v2/gmin/UpdateDialog.kt
@@ -41,7 +41,7 @@ class UpdateDialog(val context: Context){
         //수정 이벤트 처리
         binding.btnUdEnterMin.setOnClickListener{
             val edtUdContent = binding.edtUdContentMin.text.toString()
-            val edtUdPay = binding.edtUdPayMin.text.toString()
+            val edtUdPay = binding.edtUdPayMin.text.toString().toInt()
             val edtUdDate = binding.edtUdDateMin.text.toString()
             val ABitemViewData = ItemViewData(edtUdDate, edtUdContent, edtUdPay)
             (context as ExpensesActivity).updateItemViewDataList(position, ABitemViewData)
